requirements.txt
langchain_core==0.3.59
langchain_ollama==0.3.2
langgraph==0.4.3
python_fasthtml==0.12.14
